requirements.txt
asgiref==3.8.1
Django>=3.2,<4.0
sqlparse==0.5.0
djangorestframework>=3.1,<4 
